Remove noex::InstanceAdmin.
Also remove all the tests for this class, and made the last few changes
so bigtable::InstanceAdmin does not depend on it.
