+++ google/cloud/bigtable/bigtable_client.bzl
@@ -36,9 +36,6 @@ bigtable_client_hdrs = [
     "instance_update_config.h",
     "internal/async_bulk_apply.h",
     "internal/async_check_consistency.h",
-    "internal/async_list_app_profiles.h",
-    "internal/async_list_clusters.h",
-    "internal/async_list_instances.h",
     "internal/async_longrunning_op.h",
     "internal/async_loop_op.h",
     "internal/async_op_traits.h",
@@ -57,7 +54,6 @@ bigtable_client_hdrs = [
     "internal/common_client.h",
     "internal/conjunction.h",
     "internal/grpc_error_delegate.h",
-    "internal/instance_admin.h",
     "internal/poll_longrunning_operation.h",
     "internal/prefix_range_end.h",
     "internal/readrowsparser.h",
@@ -105,7 +101,6 @@ bigtable_client_srcs = [
     "internal/completion_queue_impl.cc",
     "internal/common_client.cc",
     "internal/grpc_error_delegate.cc",
-    "internal/instance_admin.cc",
     "internal/prefix_range_end.cc",
     "internal/readrowsparser.cc",
     "internal/rowreaderiterator.cc",
