+++ google/cloud/bigtable/bigtable_client_unit_tests.bzl
@@ -38,9 +38,6 @@ bigtable_client_unit_tests = [
     "instance_config_test.cc",
     "instance_update_config_test.cc",
     "internal/async_check_consistency_test.cc",
-    "internal/async_list_app_profiles_test.cc",
-    "internal/async_list_clusters_test.cc",
-    "internal/async_list_instances_test.cc",
     "internal/async_longrunning_op_test.cc",
     "internal/async_poll_op_test.cc",
     "internal/async_retry_multi_page_test.cc",
@@ -49,7 +46,6 @@ bigtable_client_unit_tests = [
     "internal/async_retry_unary_rpc_test.cc",
     "internal/bulk_mutator_test.cc",
     "internal/table_async_check_and_mutate_row_test.cc",
-    "internal/instance_admin_test.cc",
     "internal/grpc_error_delegate_test.cc",
     "internal/prefix_range_end_test.cc",
     "internal/table_admin_test.cc",
