+++ google/cloud/bigtable/tests/CMakeLists.txt
@@ -26,7 +26,6 @@ set(bigtable_client_integration_tests
     filters_integration_test.cc
     instance_admin_integration_test.cc
     instance_admin_async_future_integration_test.cc
-    instance_admin_async_integration_test.cc
     mutations_integration_test.cc)
 
 include(CreateBazelConfig)
