+++ google/cloud/bigtable/tests/bigtable_client_integration_tests.bzl
@@ -26,6 +26,5 @@ bigtable_client_integration_tests = [
     "filters_integration_test.cc",
     "instance_admin_integration_test.cc",
     "instance_admin_async_future_integration_test.cc",
-    "instance_admin_async_integration_test.cc",
     "mutations_integration_test.cc",
 ]
